# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/user/api/v1/user": {
      "get": {
        "operationId": "UserController_companyList",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortDirection",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
    "/user/api/v1/user/prifle": {
      "get": {
        "operationId": "UserController_userProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Fetch profile of login user",
        "tags": [
          "User"
        ]
      },
      "patch": {
        "operationId": "UserController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string",
                    "example": "John"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "Doe"
                  },
                  "email": {
                    "type": "string",
                    "example": "john@example.com"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "example": "+123456789"
                  },
                  "profilePicture": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update profile (with optional profile picture)",
        "tags": [
          "User"
        ]
      }
    },
    "/user/api/v1/user/{userId}": {
      "get": {
        "operationId": "UserController_getUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "User"
        ]
      },
      "delete": {
        "operationId": "UserController_archivedUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
    "/user/api/v1/user/{userId}/unarchive": {
      "patch": {
        "operationId": "UserController_unarchiveUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
    "/user/api/v1/user/profile/{userId}": {
      "patch": {
        "operationId": "UserController_updateProfilewWithUserId",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string",
                    "example": "John"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "Doe"
                  },
                  "email": {
                    "type": "string",
                    "example": "john@example.com"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "example": "+123456789"
                  },
                  "profilePicture": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update profile (with optional profile picture)",
        "tags": [
          "User"
        ]
      }
    },
    "/user/api/v1/user/employee/{userId}": {
      "patch": {
        "operationId": "UserController_updateEmployeeInfo",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/EmployeeDto"
                  },
                  {
                    "type": "object",
                    "properties": {
                      "profilePicture": {
                        "type": "string",
                        "format": "binary"
                      },
                      "passportId": {
                        "type": "string",
                        "format": "binary"
                      },
                      "proofOfAddress": {
                        "type": "string",
                        "format": "binary"
                      },
                      "otherProofOfIdentification": {
                        "type": "string",
                        "format": "binary"
                      }
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update profile (with optional profile picture)",
        "tags": [
          "User"
        ]
      }
    },
    "/user/api/v1/user/change-password": {
      "patch": {
        "operationId": "UserController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePassword"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
    "/user/api/v1/user/change-status/{userId}": {
      "patch": {
        "operationId": "UserController_changeStatus",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeStatus"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
    "/user/api/v1/user/invite": {
      "post": {
        "operationId": "UserController_inviteUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteUserDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "User"
        ]
      },
      "get": {
        "operationId": "UserController_invites",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortDirection",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
    "/user/api/v1/user/re-invite//{inviteLink}": {
      "post": {
        "operationId": "UserController_reInviteUser",
        "parameters": [
          {
            "name": "X-Company-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "inviteLink",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
    "/user/api/v1/user/profile/picture": {
      "delete": {
        "operationId": "UserController_removeProfilePic",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
...
  },
  "info": {
    "title": "DigiLog User API Documentation",
    "description": "The DigiLog API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "in": "header"
      }
    },
    "schemas": {
...
      "BankInformationDto": {
        "type": "object",
        "properties": {
          "accountNumber": {
            "type": "string",
            "description": "Bank account number"
          },
          "ifsCode": {
            "type": "string",
            "description": "IFSC / IFS code of the bank"
          },
          "bankName": {
            "type": "string",
            "description": "Bank name"
          },
          "branchName": {
            "type": "string",
            "description": "Branch name of the bank"
          }
        }
      },
      "EmployeeDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "description": "Residential address of the employee"
          },
          "phoneNumber": {
            "type": "string",
            "description": "Phone number of the employee"
          },
          "maritalStatus": {
            "type": "string",
            "enum": [
              "SINGLE",
              "MARRIED"
            ],
            "description": "Marital status of the employee"
          },
          "bio": {
            "type": "string",
            "description": "Short biography of the employee"
          },
          "interest": {
            "type": "string",
            "description": "Employee interests"
          },
          "gender": {
            "type": "string",
            "enum": [
              "MALE",
              "FEMALE"
            ],
            "description": "Gender of the employee"
          },
          "religion": {
            "type": "string",
            "description": "Religion of the employee"
          },
          "altPhoneNumber": {
            "type": "string",
            "description": "Alternative phone number"
          },
          "dateOfBirth": {
            "type": "string",
            "description": "Date of birth (string format, e.g. YYYY-MM-DD)"
          },
          "bloodGroup": {
            "type": "string",
            "enum": [
              "A_POS",
              "A_NEG",
              "B_POS",
              "B_NEG",
              "AB_POS",
              "AB_NEG",
              "O_POS",
              "O_NEG"
            ],
            "description": "Blood group of the employee"
          },
          "allergy": {
            "type": "string",
            "description": "Known allergies"
          },
          "countryCode": {
            "type": "string",
            "description": "Country dialing code"
          },
          "timezone": {
            "type": "string",
            "description": "Timezone of the employee"
          },
          "currencyCode": {
            "type": "string",
            "description": "Currency code of the employee"
          },
          "payRate": {
            "type": "number",
            "description": "Pay rate (default 0)"
          },
          "period": {
            "type": "string",
            "enum": [
              "PER_HOUR",
              "PER_DAY",
              "PER_WEEK",
              "PER_MONTH",
              "PER_YEAR"
            ],
            "description": "Billing period"
          },
          "annualLeave": {
            "type": "number",
            "description": "Annual leave days"
          },
          "bankHoliday": {
            "type": "number",
            "description": "Bank holidays"
          },
          "hours": {
            "type": "number",
            "description": "Work hours"
          },
          "breakTime": {
            "type": "string",
            "description": "Break time"
          },
          "screenshotFrequency": {
            "type": "string",
            "enum": [
              "NONE",
              "ONE_X",
              "TWO_X",
              "THREE_X"
            ],
            "description": "Screenshot frequency"
          },
          "screenshotIntervalMinutes": {
            "type": "number",
            "description": "Screenshot interval in minutes"
          },
          "appTrackingType": {
            "type": "string",
            "enum": [
              "OFF",
              "APPS_ONLY",
              "URLS_ONLY",
              "BOTH"
            ],
            "description": "App tracking type (OFF, APPS_ONLY, URLS_ONLY, BOTH)"
          },
          "appScrennshotNotification": {
            "type": "boolean",
            "description": "Notify user when screenshot is taken"
          },
          "branchIds": {
            "description": "Array of branch IDs employee belongs to",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "departmentIds": {
            "description": "Array of department IDs employee belongs to",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "jobInformation": {
            "$ref": "#/components/schemas/JobInformationDto"
          },
          "emergencyContact": {
            "$ref": "#/components/schemas/EmergencyContactDto"
          },
          "bankInformation": {
            "$ref": "#/components/schemas/BankInformationDto"
          }
        }
      },
      "ChangePassword": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          },
          "confirmPassword": {
            "type": "string"
          }
        },
        "required": [
          "password",
          "newPassword",
          "confirmPassword"
        ]
      },
      "ChangeStatus": {
        "type": "object",
        "properties": {
          "active": {
            "type": "boolean",
            "description": "Update user status to active and inactive",
            "example": true
          }
        },
        "required": [
          "active"
        ]
      },
      "InviteUserDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john.doe@example.com"
          },
          "memberId": {
            "type": "string",
            "example": "1234ABDC"
          },
          "roleId": {
            "type": "string",
            "example": "Doe"
          },
          "branchId": {
            "type": "string",
            "example": "Doe"
          }
        }
      },
...
    }
  }
}